rename the yamux package (#1452)
