+++ config/muxer_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	bhost "github.com/libp2p/go-libp2p/p2p/host/basic"
-	yamux "github.com/libp2p/go-libp2p/p2p/muxer/yamux"
+	"github.com/libp2p/go-libp2p/p2p/muxer/yamux"
 	swarmt "github.com/libp2p/go-libp2p/p2p/net/swarm/testing"
 
 	"github.com/libp2p/go-libp2p-core/host"
