+++ p2p/muxer/yamux/conn.go
@@ -1,4 +1,4 @@
-package sm_yamux
+package yamux
 
 import (
 	"context"
