+++ p2p/muxer/yamux/transport_test.go
@@ -1,4 +1,4 @@
-package sm_yamux
+package yamux
 
 import (
 	"testing"
