+++ p2p/net/swarm/dial_worker_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"time"
 
-	yamux "github.com/libp2p/go-libp2p/p2p/muxer/yamux"
+	"github.com/libp2p/go-libp2p/p2p/muxer/yamux"
 	quic "github.com/libp2p/go-libp2p/p2p/transport/quic"
 	"github.com/libp2p/go-libp2p/p2p/transport/tcp"
